Basic dockerfile
Nothing clever here, nor done using dockertest as this is a slightly
weird case.  A few bits working around some changes to the rocker
project and it looks like a few packages have changed recently (had
to install libssl-dev/unstable for git2r and libxml2-dev for xml2
but not sure what pulled these in).