New support for storing worker info and killing local workers
The two bits are linked via access to the PID without parsing worker
names.  Lots of cascading test failures while building this so don't
use in production.  Also another breaking change between worker/controller
though this one is not that critical because it's not the "core"
functionality.

This will be extended soon to store environment information.