Support for refreshing environments.
This was always possible by creating a new queue object, but now

    obj$refresh_environment()

will refresh the contents of the environment.  Updating the global
copy is currently at the whim of the `global` flag to this function,
and will remain so until I decide what to do about the global enviroment
(see issue #2).