UI improvement for observer, in case of single running queue
The queue name is optional when there is just a single queue.
Probably need to do this for the queue when reconnecting too, but it
seems more natural for the observer as that's a read-only thing and
will always (or reasonably always) be started into a queue that is
already running.